requirements.txt
pandas
openai
python-dotenv
langchain
langchain-openai
langgraph             

fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# 제거 항목:
#   streamlit  → FastAPI + HTML 구조로 전환되어 불필요
#   chromadb   → 현재 코드에서 미사용 (RAG 구현 시 복구)

# 실행 방법:
#   uvicorn server:app --reload --port 8000
